+++ app/Http/Controllers/PublishController.php
@@ -311,17 +311,22 @@ public function updateItemDeposits(Request $request, $deposit) {
 
   // Hapus file lama jika ada file baru yang diunggah
   if ($request->hasFile('fileUpload')) {
-    Storage::delete($deposit->file_upload);
+    // Pastikan file_upload memiliki nilai sebelum menghapus file lama
+    if ($deposit->file_upload) {
+        Storage::delete($deposit->file_upload);
+    }
     $file_path = $request->file('fileUpload')->store('public/fileUploads');
     $deposit->file_upload = $file_path;
   }
 
   // Hapus gambar lama jika ada gambar baru yang diunggah
   if ($request->hasFile('image')) {
-      Storage::delete($deposit->image);
-      $image = $request->file('image')->store('public/images');
-      $deposit->image = $image;
-  }
+    if ($deposit->image) {
+        Storage::delete($deposit->image);
+    }
+    $image_path = $request->file('image')->store('public/images');
+    $deposit->image = $image_path;
+ }
 
 
   $deposit->update([
